tests/wpt/tests/html/canvas/offscreen/text/2d.text.drawing.style.fontKerning.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.drawing.style.fontKerning</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.drawing.style.fontKerning</h1>
<p class="desc">Testing basic functionalities of fontKerning for canvas</p>


<script>
var t = async_test("Testing basic functionalities of fontKerning for canvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertSame(ctx.fontKerning, "auto", "ctx.fontKerning", "\"auto\"");
  ctx.fontKerning = "normal";
  _assertSame(ctx.fontKerning, "normal", "ctx.fontKerning", "\"normal\"");
  width_normal = ctx.measureText("TAWATAVA").width;
  ctx.fontKerning = "none";
  _assertSame(ctx.fontKerning, "none", "ctx.fontKerning", "\"none\"");
  width_none = ctx.measureText("TAWATAVA").width;
  _assert(width_normal < width_none, "width_normal < width_none");
  t.done();

});
</script>
