Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.fontKerning-expected.txt

2d.text.drawing.style.fontKerning

Testing basic functionalities of fontKerning for canvas


FAIL Testing basic functionalities of fontKerning for canvas assert_equals: ctx.fontKerning === "Auto" (got [undefined], expected Auto[string]) expected (string) "Auto" but got (undefined) undefined
